pyproject.toml
[project]
name = "chess-rllm"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "chess",
    "google-genai>=1.68.0",
    "python-dotenv>=1.2.2",
    "rllm[tinker]",
]

[tool.uv.sources]
rllm = { git = "https://github.com/rllm-org/rllm.git" }
